KFSPTS-21717 Initial XML converter updates for KEW
